Desktop/parishiksha-retrieval-assistant/parishiksha-retrieval-assistant/requirements.txt
pdfplumber
pymupdf
transformers
torch
rank-bm25
sentence-transformers
openai>=1.14.0
chromadb>=0.5.0
python-dotenv
streamlit
pandas
numpy
tqdm
matplotlib
seaborn
langchain-text-splitters